Разделил файл функций
